+++ changelog.md
@@ -1,2 +1,2 @@
 ### Durability Tooltip 1.1.5
-- Initial release of Durability Tooltip for Minecraft 1.21
+- Initial release of Durability Tooltip for Minecraft 1.21.6 - 1.21.8